Added moving platform asset package
Moving platform requires the developer to set the number of waypoints,
and then drag the transforms of waypoint objects onto the waypoints to
set up the references. The platform will then move from waypoint to
waypoint in sequential order waiting at each point for the indicated
amount of seconds until it reaches the end, and then will return along
the path it came.